samples/python/hosted-agents/bring-your-own/invocations/github-copilot/requirements.txt
github-copilot-sdk>=0.2.0

azure-ai-agentserver-invocations==1.0.0b2
